MOved constants to seperate file

+++ components/smart_signage/include/ctrl/ctrl_const.h
@@ -3,6 +3,6 @@
 
 namespace esphome::smart_signage::ctrl {
 
-constexpr uint8_t kDefaultSessionMins = 0;
+constexpr uint32_t kDefaultSessionMins = 0;
 
 } // namespace esphome::smart_signage::ctrl
