+++ components/smart_signage/include/led/led_const.h
@@ -3,8 +3,8 @@
 
 namespace esphome::smart_signage::led {
 
-constexpr uint8_t kMinBrightPct     = 0;
-constexpr uint8_t kMaxBrightPct     = 100;
-constexpr uint8_t kDefaultBrightPct = 100;
+constexpr uint32_t kMinBrightPct     = 0;
+constexpr uint32_t kMaxBrightPct     = 100;
+constexpr uint32_t kDefaultBrightPct = 100;
 
 } // namespace esphome::smart_signage::led
